+++ CHANGES.md
@@ -4,7 +4,7 @@
 - Set switch base to prevent compiler downgrades (#12, fixes #7, @emillon)
 - Add `create --with=features` (#14, @emillon)
 - Set OPAMCLI=2.0 in opam subprocesses (#18, fixes #15, @emillon)
-- Remove empty switch if `create` fails (#17, #18, @emillon)
+- Remove empty switch if `create` fails (#17, #18, #21, @emillon)
 
 ## 0.1.1 (2020-10-06)